Separate TaskRegistry from ModelRegistry
