+++ nvflare/client/model_registry.py
@@ -12,19 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import copy
-from typing import Dict, Optional
+from typing import Optional
 
-from nvflare.app_common.abstract.exchange_task import ExchangeTask
 from nvflare.app_common.abstract.fl_model import FLModel, ParamsType
 from nvflare.app_common.data_exchange.data_exchanger import DataExchanger
 
 from .config import ClientConfig
-from .constants import SYS_ATTRS
+from .task_registry import TaskRegistry
 from .utils import DIFF_FUNCS
 
 
-class ModelRegistry:
+class ModelRegistry(TaskRegistry):
     """This class is used to remember attributes that need to share for a user code.
 
     For example, after "global_evaluate" we should remember the "metrics" value.
@@ -37,61 +35,16 @@ class ModelRegistry:
     """
 
     def __init__(self, config: ClientConfig, rank, data_exchanger: Optional[DataExchanger] = None):
-        self.data_exchanger = data_exchanger
-        self.config = config
-
-        self.received_task: Optional[ExchangeTask] = None
-        self.task_name: str = ""
-        self.cache_loaded = False
+        super().__init__(config, rank, data_exchanger)
         self.metrics = None
-        self.sys_info = {}
-        for k, v in self.config.config.items():
-            if k in SYS_ATTRS:
-                self.sys_info[k] = v
-        self.rank = rank
-
-    def receive(self, timeout: Optional[float] = None):
-        if not self.data_exchanger:
-            return None
-
-        task_name, task = self.data_exchanger.receive_data(timeout)
-
-        if task is None:
-            raise RuntimeError("no received task")
-
-        if task.data is None:
-            raise RuntimeError("no received model")
-
-        self.received_task = task
-        self.task_name = task_name
-        self.cache_loaded = True
-
-    def get_task(self, timeout: Optional[float] = None) -> Optional[ExchangeTask]:
-        try:
-            if not self.cache_loaded:
-                self.receive()
-            return copy.deepcopy(self.received_task)
-        except:
-            return None
 
     def get_model(self, timeout: Optional[float] = None) -> Optional[FLModel]:
-        try:
-            if not self.cache_loaded:
-                self.receive()
-            if self.received_task:
-                return copy.deepcopy(self.received_task.data)
-            return None
-        except:
-            return None
-
-    def get_sys_info(self) -> Dict:
-        return self.sys_info
-
-    def send_task(self, data, meta: dict, return_code: str) -> None:
-        if not self.data_exchanger or not self.task_name or self.received_task is None:
-            return None
-        task = ExchangeTask(self.task_name, self.received_task.task_id, meta, data, return_code)
-        self.data_exchanger.submit_data(data=task)
+        task = self.get_task()
+        if task is not None and task.data is not None:
+            if not isinstance(task.data, FLModel):
+                raise RuntimeError("task.data is not FLModel.")
+            return task.data
+        return None
 
     def send_model(self, model: FLModel) -> None:
         if not self.data_exchanger:
@@ -105,6 +58,8 @@ class ModelRegistry:
                 raise RuntimeError("no received task")
             elif self.received_task.data is None:
                 raise RuntimeError("no received model")
+            elif not isinstance(self.received_task.data, FLModel):
+                raise RuntimeError("received_task.data is not FLModel.")
             elif model.params is not None:
                 if model.params_type == ParamsType.FULL:
                     try:
@@ -114,16 +69,8 @@ class ModelRegistry:
                         raise RuntimeError(f"params diff function failed: {e}")
             elif model.metrics is None:
                 raise RuntimeError("the model to send does not have either params or metrics")
-        self.data_exchanger.submit_data(data=model)
+        self.send_task(model, {}, "ok")
 
     def clear(self):
-        self.received_task = None
-        self.cache_loaded = False
+        super().clear()
         self.metrics = None
-
-    def __str__(self):
-        return f"{self.__class__.__name__}(config: {self.config.get_config()})"
-
-    def __del__(self):
-        if self.data_exchanger:
-            self.data_exchanger.finalize(close_pipe=False)
